requirements.txt
python-telegram-bot==21.6
playwright==1.47.0
